backend/pyproject.toml
[tool.poetry]
name = "Resume Match Pro Backend"
version = "0.1.0"
description = ""
authors = ["ppokrovskii"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = "^0.29.0"
sqlalchemy = "^2.0.29"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
pydantic = "^2.6.4"
python-dotenv = "^1.0.1"
sqlmodel = "^0.0.16"
databases = "^0.9.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
